requirements.txt
psycopg2==2.5.3

# Django
Django==1.6.5

South==1.0

Pillow==2.5.3

Fabric==1.9.1
fabtools==0.19.0
Jinja2==2.7.3

# Django batteries
django-grappelli==2.5.3
django-localeurl==2.0.1
python-social-auth==0.1.26
djangorestframework==2.4.0